Add close command for threads/posts
The command will close the thread (if on a text channel),
and archive the post (if on a forum channel).

When in a forum channel, the command will look for a tag
from that specific forum called 'done' (each tag will have
different IDs depending on the forum) and apply to the post
if found.

The bot will send a message notifying who closed it.