inbox support v1 (#28)
Added support for marking messages with an inbox emoji to keep track of them.

Typical use case would be to use the inbox emoji to signal to other people on discord that you saw a message, but you either will handle it later, or are already handling it async but it will take more time.

Having a bot keep track of all those messages will make it easier to keep track of them :)

the !inbox command, just like all the other ones, can be also issued on a DM with the bot.