admin speedup: defer big json fields to optimise the admin list page (#38)
The list of objects in django admin doesn't use the big jsons fields, so it's okay to defer them. This makes for a significant speed up as some of the JSONs are a few MB in size, and loading a list of 50 or a 100 (one page) of those objects would mean 100s of MB more data to read from the database that is then not even used.